finished

+++ app/Http/Controllers/ProductController.php
@@ -8,26 +8,25 @@ use App\Models\Products;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Str;
 
 class ProductController extends Controller
 {
 
     public function index()
     {
-        $products = DB::table('products')->orderBy('id', 'desc')->cursorPaginate(10,'id');
+        $products = DB::table('products')->orderBy('id', 'desc')->cursorPaginate(10);
         return response()->json(['products' => $products]);
     }
     public function getProduct(Products $product)
     {
        $product->load('owner_details');
-       return response()->json(['product' => $product]);
+       return response()->json( $product);
     }
     public function createProduct(ProductCreate $request)
     {
         $image = $request->file('image');
-        $image_name = Str::uniqid() .'.'. $image->extension();
-        $image->storePubliclyAs('/products/images', $image_name);
+        $image_name = uniqid() .'.'. $image->extension();
+        $image->storeAs('/public/products/images', $image_name);
 
         $product = Products::create([
             'title' => $request->title,
@@ -36,30 +35,33 @@ class ProductController extends Controller
             'owner' => $request->user()->id,
             'image' => asset('/storage/products/images/'. $image_name),
         ]);
-        return response()->json(['product' => $product], 201);
+        return response()->json($product, 201);
     }
 
     public function updateProduct(ProductUpdate $request, Products $product)
     {
+        $product->load('owner_details');
         $product->title = $request->title ?? $product->title;
-        $product->description == $request->description ?? $product->description;
-        $product->price == $request->price ?? $product->price;
+        $product->description = $request->description ?? $product->description;
+        $product->price = $request->price ?? $product->price;
 
         //image update
         $image = $request->file('image');
         if ($image) {
-            $image_name = Str::uniqid() .'.'. $image->extension();
+            $image_name = uniqid() .'.'. $image->extension();
             
             //delte old file
             $location = strstr($product->image,'products');
-            Storage::disk('public')->delete($location);
+            if ($location !== 'products/images/default.jpg') {
+                Storage::disk('public')->delete($location);
+            }
             
             //save new file
-            $image->storePubliclyAs('/products/images', $image_name);
+            $image->storeAs('/public/products/images', $image_name);
             $product->image = asset('/storage/products/images/'. $image_name);
         }
         $product->save();
-        return response()->json(['product' => $product]);
+        return response()->json($product, 200);
     }
 
     public function deleteProduct(Request $request, Products $product)
